zb-8: Support slots without swap status
Added support for slots that do not contain a swap status. In this case
there is no fallback if the swap is interrupted but it saves flash as
there is no storage of the status.

Slots that do not have a swap status are intended mainly for data slots,
but if there is insufficient space it can also be used for images.

Signed-off-by: Laczen JMS <laczenjms@gmail.com>